requirements.txt
Django>=4.2
djangorestframework
psycopg2-binary
Pillow
django-widget-tweaks
django-crispy-forms
crispy-bootstrap5
python-dotenv
